SCB-327 minor refactor,delete metrics-common and move all content to foundation-metrics
