+++ demo/perf/src/main/java/org/apache/servicecomb/demo/perf/PerfMetricsFilePublisher.java
@@ -19,10 +19,10 @@ package org.apache.servicecomb.demo.perf;
 import java.util.Map;
 import java.util.Map.Entry;
 
+import org.apache.servicecomb.foundation.metrics.MetricsConst;
+import org.apache.servicecomb.foundation.metrics.publish.MetricNode;
+import org.apache.servicecomb.foundation.metrics.publish.MetricsLoader;
 import org.apache.servicecomb.foundation.vertx.VertxUtils;
-import org.apache.servicecomb.metrics.common.MetricsConst;
-import org.apache.servicecomb.metrics.common.publish.MetricNode;
-import org.apache.servicecomb.metrics.common.publish.MetricsLoader;
 import org.apache.servicecomb.metrics.core.publish.DataSource;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
