+++ foundations/foundation-metrics/src/main/java/org/apache/servicecomb/foundation/metrics/publish/HealthCheckResult.java
@@ -15,12 +15,42 @@
  * limitations under the License.
  */
 
-package org.apache.servicecomb.metrics.common;
+package org.apache.servicecomb.foundation.metrics.publish;
 
-import java.util.Map;
+public class HealthCheckResult {
+  private boolean healthy;
 
-public interface HealthCheckerPublisher {
-  Map<String, HealthCheckResult> health();
+  private String information;
 
-  HealthCheckResult healthWithName(String name);
+  //unsupport object or generic type,so string..
+  private String extraData;
+
+  private long timestamp;
+
+  public boolean isHealthy() {
+    return healthy;
+  }
+
+  public String getInformation() {
+    return information;
+  }
+
+  public String getExtraData() {
+    return extraData;
+  }
+
+  public long getTimestamp() {
+    return timestamp;
+  }
+
+  public HealthCheckResult() {
+  }
+
+  public HealthCheckResult(boolean healthy, String information, String extraData) {
+    this();
+    this.healthy = healthy;
+    this.information = information;
+    this.extraData = extraData;
+    this.timestamp = System.currentTimeMillis();
+  }
 }
