+++ foundations/foundation-metrics/src/main/java/org/apache/servicecomb/foundation/metrics/publish/MetricsLoader.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.servicecomb.metrics.common.publish;
+package org.apache.servicecomb.foundation.metrics.publish;
 
 import java.util.ArrayList;
 import java.util.HashMap;
