+++ metrics/pom.xml
@@ -32,9 +32,9 @@
   <name>Java Chassis::Metrics</name>
 
   <modules>
-    <module>metrics-common</module>
     <module>metrics-core</module>
     <module>metrics-extension</module>
     <module>metrics-integration</module>
   </modules>
+
 </project>
